fix: modulize the bedrock invocation
