+++ .github/workflows/code-review.yml
@@ -56,7 +56,7 @@ jobs:
         aws-region: us-east-1
 
     - name: Intelligent GitHub Actions
-      uses: yike5460/intelli-ops@v0.0.82
+      uses: yike5460/intelli-ops@v0.0.83
       with:
         # Automatic Provision: The GITHUB_TOKEN is automatically created and provided by GitHub for each workflow run. You don't need to manually create or store this token as a secret.
         github-token: ${{ secrets.GITHUB_TOKEN }}
